Remove the code to perform incremental backup of the file repository
This code was put in place because a naive backup of the file repository
was impossible already for reasonably size projects. The problem was the
underlying design where each node had an associated folder on disk, even
if it contained no files, and all files were stored separately, creating
a huge numbers of files. This meant that just rsync'ing the folder could
take days even to just calculate the difference with the backed up
folder.

The ad-hoc solution was this script that would only transfer the folders
of the nodes added since the last backup date, the list of which was
determined by a simple query. However, now that the new file repository
is implemented, which has been explicitly designed to be easily backed
up by tools like `rsync`, this custom solution is no longer needed.