🐛 FIX: Add `broker_*` parameters to default test config. (#4831)
In the `TestManager`, set the `broker_*` parameters by adding
them to the `_DEFAULT_PROFILE_INFO`.

The default values were copied from the config schema. They could
be extracted directly, but since some keys (e.g. `AIIDADB_*`)
are named differently in the `profile_info` compared to the
config schema, this may be unreliable.

Previously, these were set to `None` by default, which caused the
config validation to fail because only string values are allowed.